01-install-eks-cluster/requirements.txt
aws-cdk.core==1.27.0
aws-cdk.aws-eks==1.27.0
aws-cdk.aws-ec2==1.27.0
aws-cdk.aws-iam==1.27.0
aws-cdk.aws-ecr==1.27.0
pyyaml==5.2
